[ADD/#26] 일기 삭제하기 바텀시트 파일 생성
