+++ app/src/main/java/com/sopt/clody/presentation/ui/diarylist/component/DailyDiaryCard.kt
@@ -35,6 +35,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import com.sopt.clody.R
 import com.sopt.clody.presentation.ui.component.bottomsheet.ClodyBottomSheet
+import com.sopt.clody.presentation.ui.component.bottomsheet.DiaryDeleteSheet
 import com.sopt.clody.presentation.ui.component.dialog.ClodyDialog
 import com.sopt.clody.ui.theme.ClodyTheme
 
@@ -147,59 +148,3 @@ fun DailyDiaryCard(
         )
     }
 }
-
-@Composable
-fun DiaryDeleteSheet(
-    onDismiss: () -> Unit,
-    onShowDiaryDeleteDialogStateChange: (Boolean) -> Unit
-) {
-    ClodyBottomSheet(
-        onDismissRequest = onDismiss,
-        content = {
-            DiaryDeleteBottomSheetItem(
-                onDismiss = onDismiss,
-                onShowDiaryDeleteDialogStateChange = onShowDiaryDeleteDialogStateChange
-            )
-        }
-    )
-}
-
-@Composable
-fun DiaryDeleteBottomSheetItem(
-    onDismiss: () -> Unit,
-    onShowDiaryDeleteDialogStateChange: (Boolean) -> Unit,
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .background(ClodyTheme.colors.white)
-            .padding(vertical = 20.dp),
-        verticalArrangement = Arrangement.Top
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier
-                .fillMaxWidth()
-                .clip(RoundedCornerShape(10.dp))
-                .clickable {
-                    onDismiss()
-                    onShowDiaryDeleteDialogStateChange(true)
-                }
-                .padding(start = 24.dp)
-        ) {
-            Image(
-                painter = painterResource(id = R.drawable.ic_bottomsheet_trash),
-                contentDescription = null
-            )
-            Spacer(modifier = Modifier.width(8.dp))
-            Text(
-                text = "삭제하기",
-                style = ClodyTheme.typography.body4SemiBold,
-                color = ClodyTheme.colors.gray01
-            )
-        }
-        Spacer(modifier = Modifier.navigationBarsPadding())
-        Spacer(modifier = Modifier.height(60.dp))
-    }
-}
-
